Use updated neovim config setup
Neovim used to suggest this symlink between an init file and an existing
vimrc as a way to transition from vim to neovim. Now there's a different
recommended way of doing the same thing. This change updates to using
that new method and makes use of the more flexible install task to do
it.

With the old install setup, this would symlink an entire ~/.config
directory just for a single file. Since a handful of other apps use that
directory and I don't have a great reason to add all of their configs to
source control, naming the init.vim file a little differently and adding
a custom step to symlink it to its proper location works best.