Fix macOS tmux-256color term
Using this term lets tmux render fonts with more variations (e.g.
italics). But something about it is broken on macOS when you try to run
certain terminal commands. There's a warning about the terminal not
being fully functional. There's an issue on the tmux repo that outlines
the issue and the fix: https://github.com/tmux/tmux/issues/2262

This downloads the tmux-256color term linked by the tmux contributor on
that issue, checks that the contents haven't changed since a known good
version, and installs it properly.