Customize sonokai theme for tree-sitter
Sonokai already supports tree-sitter directly, but it has a couple of
slightly strange highlights.

This makes Ruby class and instance variables (TSLabel) orange italic
(the default without tree-sitter) instead of the red that blends in with
`def`, `end`, and friends.

It also makes it so that constants, which include the colon in kwargs or
a 1.9-style hash with symbol keys, don't have a funky italic style to
them, and it further sets the class and inst variables apart.