Delcare (near) PATH bankruptcy
The key here is putting rbenv init at the bottom of this file so rbenv
shims are at the very front of the PATH. But there's also a ton in my
PATH that I don't think I really need and can't quite remember where it
came from. Getting rid of everything I don't _know_ that I need and
we'll see what happens.