Introduce cmp and snippy for neovim
These are essentially replacing supertab and snipmate. I went with
snippy because it is lua but also uses the same stock snippets as
snipmate, which makes for a future-looking but easier transition.

cmp has a ton going for it. For now, I've just started with the
supertab-like setup, and I'll add from here.