Fix vim-tmux terminal colors on linux
I'm not sure why this works, but it does. Without it the vim colorscheme
within tmux was slightly off (i.e. tinted brown), which also changed the
tmux colorscheme to match. Vim outside of tmux was fine. This line fixes
it, so I'm leaving it and making it linux-specific for now.