Install macOS apps
This is more about automating the initial setup of a new computer than
getting a dev environment up and running, but it's nice to have
everything in one place.

In the future, I'd love to make more of a mac/linux separation and maybe
even a main/remote machine distinction. That gets complicated quickly,
so I'm trying to stick to most of what's common among those concerns.
This changes that a bit, but we'll see how it goes.