Add support for config directories
A variety of apps use ~/.config for their configs. I've had a special
case for .config/nvim/init.vim, but now that I want more nvim confis,
it's time to move to a more general solution.

This skips everything in config/ for the initial install pass. Then it
goes through everything in config/ one at a time, which leaves any
existing local configs in place that don't correspond to configs in this
repo.