Frequent tmux status bar refreshes
With the git branch in my status line, it can be misleading at best when
switching branches frequently to only get a refresh every 15 seconds.
And generally, I've just not relied on that bit of the status line.

I tried a few different intervals, and there are enough scenarios where
I want the near-immediate feedback that only 1 second can give. The
other bits in the status line aren't particularly intensive
calculations, so this shouldn't cause any real problems. But we'll see!