Use git as Mackup storage engine
iCloud is a pain to get to sync quickly, completely, and reliably. I
don't want the bloat of those other options, so back to trusty ole git.
The configs will live in a separate, private repo that I'll clone and
use for syncing. It's a more manual process, but I also don't expect
these configs to change all that much.