Extract & refine Ruby linting & formatting in vim
There are so many options for linting and formatting in Ruby projects,
and they vary from project to project. This extraction is the beginning
of a pattern where I'll have lang configs that won't so much correspond
to a particular file but will detect whether or not a given language is
part of a project based on files in the current working directory, so
it'll turn on the right LSP configs–and whatever else–if this is a "Ruby
project", but a Ruby project can also be a Javascript project with other
tools enabled, as well.

The key here is detecting what linting & formatting tools are in use in
the Ruby project and whether they're using a version that
supports/includes LSP. If it does, we'll use the LSP. If it doesn't,
we'll fall back to Ale. Worth noting that because Standard uses RuboCop,
they're exclusive from each other, while the others are additive. In
practice this means that a project with RuboCop and SyntaxTree will use
both formatters. The ordering is setup so that even with Ale, RuboCop
runs first so that SyntaxTree (the more opinionated formatter) can clean
up anything it needs to.

Because these lang configs rely on certain plugins, I'm enumerating
those at the top of the file (since not all–e.g. ale–involve requiring
the plugin explicitly). This will make it easier to see where a plugin
is used and the difference between general use/config of a plugin vs.
language/project-specific use.

One final note on lspconfg vs. Ale: more recent versions of Ale will
actually detect what is running under lspconfig and disable those
linters/fixers in Ale automatically. So, even though the default for Ale
is to include all the linters and fixers it can find for a language
(e.g. RuboCop), if the RuboCop LSP is running via lspconfig, Ale will
disable its own RuboCop integration. This makes things in this setup
much simpler!