Replace git-hook ctags w/ vim-gutentags
This should be a much simpler setup, keeps tags out of the .git
directory (default location is already ignored globally), will rebuild
tags selectively as I edit files in a project, and keeps configs in a
single file available for all ctags invocations. The only real downside
is the tags are only ever updated from within vim, but the trade-offs
seem well worth it at this point.